Some more compilation fixes for Windows

+++ src/core/allocator/dbc/Buddy.cpp
@@ -16,7 +16,7 @@ off_t Buddy::getFromList(uint8_t i)
 
 void Buddy::putToList(off_t addr, uint8_t i)
 {
-    REQUIRES(addr >= 0 && addr < size_);
+    REQUIRES(addr >= 0 && size_t(addr) < size_);
     REQUIRES(i >= 0);
     return __impl::core::allocator::Buddy::putToList(addr, i);
 }
